recruitment_task/requirements.txt
Django==3.2.8
django-mysql==3.8.1
django-cors-headers==3.10.0
djangorestframework==3.12.4
cryptography==35.0.0
factory-boy==3.2.0
Faker==9.3.0
PyMySQL==1.0.2
pytest==6.2.3
pytest-django==4.4.0
